# Enable kernel & memcpy overlapping in IVF index building