--- a/cpp/include/cuvs/neighbors/ivf_flat.hpp
+++ b/cpp/include/cuvs/neighbors/ivf_flat.hpp
@@ -295,6 +295,10 @@ struct index : cuvs::neighbors::index {
 /**
  * @brief Build the index from the dataset for efficient search.
  *
+ * Note, if dataset is in host memory and index_params.add_data_on_build = true, user could set a
+ * stream pool in the input raft::resource with (at least) one stream to enable kernel/copy
+ * overlapping.
+ *
  * Usage example:
  * @code{.cpp}
  *   using namespace cuvs::neighbors;
@@ -317,6 +321,10 @@ auto build(raft::resources const& handle,
 
 /**
  * @brief Build the index from the dataset for efficient search.
+ *
+ * Note, if dataset is in host memory and index_params.add_data_on_build = true, user could set a
+ * stream pool in the input raft::resource with (at least) one stream to enable kernel/copy
+ * overlapping.
  *
  * Usage example:
  * @code{.cpp}
@@ -342,6 +350,10 @@ void build(raft::resources const& handle,
 /**
  * @brief Build the index from the dataset for efficient search.
  *
+ * Note, if dataset is in host memory and index_params.add_data_on_build = true, user could set a
+ * stream pool in the input raft::resource with (at least) one stream to enable kernel/copy
+ * overlapping.
+ *
  * Usage example:
  * @code{.cpp}
  *   using namespace cuvs::neighbors;
@@ -364,6 +376,10 @@ auto build(raft::resources const& handle,
 
 /**
  * @brief Build the index from the dataset for efficient search.
+ *
+ * Note, if dataset is in host memory and index_params.add_data_on_build = true, user could set a
+ * stream pool in the input raft::resource with (at least) one stream to enable kernel/copy
+ * overlapping.
  *
  * Usage example:
  * @code{.cpp}
@@ -389,6 +405,10 @@ void build(raft::resources const& handle,
 /**
  * @brief Build the index from the dataset for efficient search.
  *
+ * Note, if dataset is in host memory and index_params.add_data_on_build = true, user could set a
+ * stream pool in the input raft::resource with (at least) one stream to enable kernel/copy
+ * overlapping.
+ *
  * Usage example:
  * @code{.cpp}
  *   using namespace cuvs::neighbors;
@@ -411,6 +431,10 @@ auto build(raft::resources const& handle,
 
 /**
  * @brief Build the index from the dataset for efficient search.
+ *
+ * Note, if dataset is in host memory and index_params.add_data_on_build = true, user could set a
+ * stream pool in the input raft::resource with (at least) one stream to enable kernel/copy
+ * overlapping.
  *
  * Usage example:
  * @code{.cpp}
@@ -436,6 +460,10 @@ void build(raft::resources const& handle,
 /**
  * @brief Build the index from the dataset for efficient search.
  *
+ * Note, if dataset is in host memory and index_params.add_data_on_build = true, user could set a
+ * stream pool in the input raft::resource with (at least) one stream to enable kernel/copy
+ * overlapping.
+ *
  * Usage example:
  * @code{.cpp}
  *   using namespace cuvs::neighbors;
@@ -458,6 +486,10 @@ auto build(raft::resources const& handle,
 
 /**
  * @brief Build the index from the dataset for efficient search.
+ *
+ * Note, if dataset is in host memory and index_params.add_data_on_build = true, user could set a
+ * stream pool in the input raft::resource with (at least) one stream to enable kernel/copy
+ * overlapping.
  *
  * Usage example:
  * @code{.cpp}
@@ -483,6 +515,10 @@ void build(raft::resources const& handle,
 /**
  * @brief Build the index from the dataset for efficient search.
  *
+ * Note, if dataset is in host memory and index_params.add_data_on_build = true, user could set a
+ * stream pool in the input raft::resource with (at least) one stream to enable kernel/copy
+ * overlapping.
+ *
  * Usage example:
  * @code{.cpp}
  *   using namespace cuvs::neighbors;
@@ -505,6 +541,10 @@ auto build(raft::resources const& handle,
 
 /**
  * @brief Build the index from the dataset for efficient search.
+ *
+ * Note, if dataset is in host memory and index_params.add_data_on_build = true, user could set a
+ * stream pool in the input raft::resource with (at least) one stream to enable kernel/copy
+ * overlapping.
  *
  * Usage example:
  * @code{.cpp}
@@ -530,6 +570,10 @@ void build(raft::resources const& handle,
 /**
  * @brief Build the index from the dataset for efficient search.
  *
+ * Note, if dataset is in host memory and index_params.add_data_on_build = true, user could set a
+ * stream pool in the input raft::resource with (at least) one stream to enable kernel/copy
+ * overlapping.
+ *
  * Usage example:
  * @code{.cpp}
  *   using namespace cuvs::neighbors;
@@ -552,6 +596,10 @@ auto build(raft::resources const& handle,
 
 /**
  * @brief Build the index from the dataset for efficient search.
+ *
+ * Note, if dataset is in host memory and index_params.add_data_on_build = true, user could set a
+ * stream pool in the input raft::resource with (at least) one stream to enable kernel/copy
+ * overlapping.
  *
  * Usage example:
  * @code{.cpp}
@@ -584,6 +632,9 @@ void build(raft::resources const& handle,
 
 /**
  * @brief Build a new index containing the data of the original plus new extra vectors.
+ *
+ * Note, if dataset is in host memory, user could set a stream pool in the input raft::resource with
+ * (at least) one stream to enable kernel/copy overlapping.
  *
  * Implementation note:
  *    The new data is clustered according to existing kmeans clusters, then the cluster
@@ -620,6 +671,9 @@ auto extend(raft::resources const& handle,
 /**
  * @brief Extend the index in-place with the new data.
  *
+ * Note, if dataset is in host memory, user could set a stream pool in the input raft::resource with
+ * (at least) one stream to enable kernel/copy overlapping.
+ *
  * Usage example:
  * @code{.cpp}
  *   using namespace cuvs::neighbors;
@@ -648,6 +702,9 @@ void extend(raft::resources const& handle,
 
 /**
  * @brief Build a new index containing the data of the original plus new extra vectors.
+ *
+ * Note, if dataset is in host memory, user could set a stream pool in the input raft::resource with
+ * (at least) one stream to enable kernel/copy overlapping.
  *
  * Implementation note:
  *    The new data is clustered according to existing kmeans clusters, then the cluster
@@ -684,6 +741,9 @@ auto extend(raft::resources const& handle,
 /**
  * @brief Extend the index in-place with the new data.
  *
+ * Note, if dataset is in host memory, user could set a stream pool in the input raft::resource with
+ * (at least) one stream to enable kernel/copy overlapping.
+ *
  * Usage example:
  * @code{.cpp}
  *   using namespace cuvs::neighbors;
@@ -712,6 +772,9 @@ void extend(raft::resources const& handle,
 
 /**
  * @brief Build a new index containing the data of the original plus new extra vectors.
+ *
+ * Note, if dataset is in host memory, user could set a stream pool in the input raft::resource with
+ * (at least) one stream to enable kernel/copy overlapping.
  *
  * Implementation note:
  *    The new data is clustered according to existing kmeans clusters, then the cluster
@@ -748,6 +811,9 @@ auto extend(raft::resources const& handle,
 /**
  * @brief Extend the index in-place with the new data.
  *
+ * Note, if dataset is in host memory, user could set a stream pool in the input raft::resource with
+ * (at least) one stream to enable kernel/copy overlapping.
+ *
  * Usage example:
  * @code{.cpp}
  *   using namespace cuvs::neighbors;
@@ -776,6 +842,9 @@ void extend(raft::resources const& handle,
 
 /**
  * @brief Build a new index containing the data of the original plus new extra vectors.
+ *
+ * Note, if dataset is in host memory, user could set a stream pool in the input raft::resource with
+ * (at least) one stream to enable kernel/copy overlapping.
  *
  * Implementation note:
  *    The new data is clustered according to existing kmeans clusters, then the cluster
@@ -812,6 +881,9 @@ auto extend(raft::resources const& handle,
 /**
  * @brief Extend the index in-place with the new data.
  *
+ * Note, if dataset is in host memory, user could set a stream pool in the input raft::resource with
+ * (at least) one stream to enable kernel/copy overlapping.
+ *
  * Usage example:
  * @code{.cpp}
  *   using namespace cuvs::neighbors;
@@ -840,6 +912,9 @@ void extend(raft::resources const& handle,
 
 /**
  * @brief Build a new index containing the data of the original plus new extra vectors.
+ *
+ * Note, if dataset is in host memory, user could set a stream pool in the input raft::resource with
+ * (at least) one stream to enable kernel/copy overlapping.
  *
  * Implementation note:
  *    The new data is clustered according to existing kmeans clusters, then the cluster
@@ -876,6 +951,9 @@ auto extend(raft::resources const& handle,
 /**
  * @brief Extend the index in-place with the new data.
  *
+ * Note, if dataset is in host memory, user could set a stream pool in the input raft::resource with
+ * (at least) one stream to enable kernel/copy overlapping.
+ *
  * Usage example:
  * @code{.cpp}
  *   using namespace cuvs::neighbors;
@@ -904,6 +982,9 @@ void extend(raft::resources const& handle,
 
 /**
  * @brief Build a new index containing the data of the original plus new extra vectors.
+ *
+ * Note, if dataset is in host memory, user could set a stream pool in the input raft::resource with
+ * (at least) one stream to enable kernel/copy overlapping.
  *
  * Implementation note:
  *    The new data is clustered according to existing kmeans clusters, then the cluster
@@ -939,6 +1020,9 @@ auto extend(raft::resources const& handle,
 
 /**
  * @brief Extend the index in-place with the new data.
+ *
+ * Note, if dataset is in host memory, user could set a stream pool in the input raft::resource with
+ * (at least) one stream to enable kernel/copy overlapping.
  *
  * Usage example:
  * @code{.cpp}