--- a/cpp/src/neighbors/ivf_flat/ivf_flat_build.cuh
+++ b/cpp/src/neighbors/ivf_flat/ivf_flat_build.cuh
@@ -196,13 +196,14 @@ void extend(raft::resources const& handle,
   constexpr size_t kReasonableMaxBatchSize = 65536;
   size_t max_batch_size                    = std::min<size_t>(n_rows, kReasonableMaxBatchSize);
 
-  // Determine if a stream pool is setup and make sure there is at least one stream in it so we
+  // Determine if a stream pool exist and make sure there is at least one stream in it so we
   // could use the stream for kernel/copy overlapping.
   auto copy_stream = raft::resource::get_cuda_stream(handle);  // Using the main stream by default
-  if (raft::resource::get_stream_pool_size(handle) >= 1) {
-    copy_stream = raft::resource::get_stream_from_stream_pool(handle);
+  if (handle.has_resource_factory(raft::resource::resource_type::CUDA_STREAM_POOL)) {
+    if (raft::resource::get_stream_pool_size(handle) >= 1) {
+      copy_stream = raft::resource::get_stream_from_stream_pool(handle);
+    }
   }
-
   // Predict the cluster labels for the new data, in batches if necessary
   bool enable_prefetch = false;
   utils::batch_load_iterator<T> vec_batches(new_vectors,