@abc99lr 2024-07-22
> > currently, user need to create a stream pool [...] so we could achieve kernel copy overlapping. But this model is not intuitive. [...]  but for end-user, do we want to add a flag to our API?
> 
> This is a good question. In other words, to enable overlapping of clustering and H2D copies, we have to do the following:
> 
> ```python
> handle = DeviceResources(n_streams=1)
> index = ivf_pq.build(build_params, dataset, resources=handle)
> ```
> 
> The question is, whether we would like to have an extra build option like `overlap_h2d_copies`, and spawn a extra stream in the background if that is set:
> 
> ```python
> build_params.overlap_hd2_copies = True
> index = ivf_pq.build(build_params, dataset)
> ```
> 
> I would prefer that we use the first option, if we have the resources (stream pool) then we use that automatically. But we would need to update the docstring of `build`, so that the user is aware that we need a handle with a stream pool for the fastest build. What do you think @cjnolet?

Hi @cjnolet , could you comment on the two choices Tamas summarized here?