@abc99lr 2024-07-27
Re-requesting reviews. 

Added documentations. Exposed the overlapping option to users. Users need to setup a stream pool in resource and pass that to the `build` or `extend` APIs.