@cjnolet 2024-07-29
@abc99lr just to add my response here quickly- I do agree with Tamas' suggestion that we should utilize the infrastructure already available in the handle to inform the user-configurable behavior of algorithms as much as possible.

That being said, the expectations of these configurations do need to be communicated more clearly to the users through the docs, I +1 to improving the docstring.