@abc99lr 2024-07-31
> @abc99lr there's some errors in the style checker. Can you please install and use pre-commit to guarantee that each commit will pass the style checker? It takes only a few seconds to install and it's automatic each time you commit. Instructions [here](https://docs.rapids.ai/api/cuvs/nightly/contributing/#python-pre-commit-hooks).

Thanks, my pre-commit was installed in my previous conda env. I had to rebuild an env so forgot to add it. Thanks for the instruction. I think we should be good now.